refactor(componets): clean up drag logic & refactor into one component
I replaced my previous drag-counter / pointer-events hybrid approach for
keeping track of drag state with a much simpler, cleaner approach based
solely on pointer-events.

In doing so I'm losing support for IE10 and earlier, but I think that's
a fair tradeoff, since I don't really have a reason to support IE10
right now.  If I change my mind in the future, I can implement support
via a pointer-events polyfill or a modernizr-based fallback.

Given that the pointer-events-based approach necessitates putting
the click handler on the same level as the class which disables pointer
events on child elements, it now makes the most sense to also
consolidate everything into a single component.  I don't think this is a
bad thing, because the functionality is all closely related and it
doesn't notably clutter up the file-input component.