fix(drophandler): handle IE11 bug wherein change is fired twice
In Internet Explorer 11, if you clear a file input programmatically,
a change event is fired.  This is in contrast to other major browsers,
which don't do this.

This affected the file-input component in that it caused the change event to
fire a second time and clear out the user's selection.  I'm handling
this by checking to ensure that at least one file was selected before
calling selectFiles.  For our purposes, we only care about the case
where the user used the file input dialog to select a file.