test(tests): add tests for onChange handler
I've added some unit tests for the onChange handler.  It's as close as
possible to being "testable", but the fact that we can't
programmatically set the value of a file input prevents it from being
truly testable.  I understand that that's necessary from a security
perspective, but it prevents us from being able to fully unit test
such things.